Facebook Authentication Done
Successfully completed the facebook linking in the app using FB API.
Permission for Public Profile Access from the user.
Used Async/Await Syntax for promise handling.
Token returned to be saved in AsyncStorage.
Skipped the Auth Screen if already logged in.
And display of Welcome Tutorial only if using app for the first time.